UDF: Fix using constant index in array
By default in UDF integers are defined as int64, but arrays indexes
(strides) are int32. So now we coerce the index value to fix a
ValueError.

Also found that output iron arrays are not zeroed. So if some values
are not set in the function, then they will be undefined. This makes
it hard to compare against numpy, in the tests, so I've zeroed them
in the functions with "out[i] = 0".